custom_exporter/requirements.txt
mysql-connector-python
prometheus_client
requests
